Renamed the files for plotting acceptance ratios
