+++ Plotting/Plotting_acceptance_ratios.h
